backend/back_store/all_requirements.txt
aiogram==2.21
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
config==0.5.1
decorator==5.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
emoji==1.7.0
environs==9.5.0
exceptiongroup==1.0.2
executing==0.8.3
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.14.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
Markdown==3.4.1
marshmallow==3.17.0
matplotlib-inline==0.1.3
multidict==6.0.2
outcome==1.2.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.29
psycopg2==2.9.5
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.6.0
python-dotenv==0.20.0
pytils==0.4.1
pytz==2022.1
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
sqlparse==0.4.3
stack-data==0.3.0
traitlets==5.3.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.6
urllib3==1.26.10
wcwidth==0.2.5
wsproto==1.2.0
yarl==1.7.2
